% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_aic}
\alias{calc_aic}
\title{Calculate AIC values for a model}
\usage{
calc_aic(k, ll)
}
\arguments{
\item{k}{integer The number of parameters}

\item{ll}{double The maximum log-likelihood (MLE) of the fitted model}
}
\description{
Calulcates the AIC values for a given model
}
\concept{utility}
