% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_ks}
\alias{calc_ks}
\title{Run a KS test on the data}
\usage{
calc_ks(x, param_est, model_type, pvalue = TRUE)
}
\arguments{
\item{x}{doubleArray Foraging distance data}

\item{param_est}{doubleArray The parameter estimates for a model}

\item{model_type}{characterArray the name of the model ran}

\item{pvalue}{Bool To return the p value (TRUE) or D statistic (FALSE).
Defaults to TRUE.}
}
\description{
Calulctate the ks test stat and p value for a given model
}
\concept{utility}
