% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_ks_boot}
\alias{calc_ks_boot}
\title{Run a bootstapped KS test on the data}
\usage{
calc_ks_boot(x, param_est, model_type)
}
\arguments{
\item{x}{doubleArray Foraging distance data}

\item{param_est}{doubleArray The parameter estimates for a model}

\item{model_type}{characterArray the name of the model ran}
}
\description{
Calulctate the ks test stat and p value for a given model
}
\concept{utility}
