% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_var_likelihood_collective}
\alias{calc_var_likelihood_collective}
\title{Calculate likelihoods for a specified parameter for the collective model}
\usage{
calc_var_likelihood_collective(
  data,
  var,
  params,
  n = 1000,
  upper = 5,
  bounds = NULL
)
}
\arguments{
\item{data}{doubleArray The data to check the parameters against. This should
be the data the parameters were fit to.}

\item{var}{characterArray The name of the variable you want to examine the
likelihood space for.}

\item{n}{integer The number of points to sample, defaults to 1000.}

\item{upper}{integer The upper bound}

\item{p}{double The proportion of scouts.}

\item{bs}{double The scout rate.}

\item{as}{double The scout alpha.}

\item{br}{double The recruit rate.}

\item{ar}{double The recruit alpha.}
}
\value{
tibble The varied parameter and associated likelihood and
log-likelihood scores.
}
\description{
Calculate likelihoods for a specified parameter to show the
likelihood space for that parameter. This uses the bounds defined for that
parameter and gets the specified number of values to show the likelihood
space.
}
\concept{utility}
