% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccdf_model_all}
\alias{ccdf_model_all}
\title{Model ccdf function for scout and recruit superposition. Stores results in
given array (y)}
\usage{
ccdf_model_all(x, y, p, ls, ln, qn, a)
}
\arguments{
\item{x}{NumericVector foraging distances}

\item{y}{NumericVector storage array for the results}

\item{p}{double Proportion of scouts (0<=p<=1)}

\item{ls}{double Scout rate}

\item{ln}{double Recruit rate}

\item{qn}{double Quality}

\item{a}{double alpha value}
}
\description{
Model ccdf function for scout and recruit superposition. Stores results in
given array (y)
}
