% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccdf_model_recruit}
\alias{ccdf_model_recruit}
\title{Model ccdf function for recruit model. Stores results in given array (y)}
\usage{
ccdf_model_recruit(x, y, ln, qn, a)
}
\arguments{
\item{x}{NumericVector foraging distances}

\item{y}{NumericVector storage array for the results}

\item{ln}{double Recruit rate}

\item{qn}{double Quality}

\item{a}{double alpha value}
}
\description{
Model ccdf function for recruit model. Stores results in given array (y)
}
