% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccdf_model_scout}
\alias{ccdf_model_scout}
\title{Model ccdf function for scout model. Stores results in given array (y)}
\usage{
ccdf_model_scout(x, y, ls, bs, as)
}
\arguments{
\item{x}{NumericVector foraging distances}

\item{y}{NumericVector storage array for the results}

\item{bs}{double Scout rate}

\item{as}{double scout alpha}
}
\description{
Model ccdf function for scout model. Stores results in given array (y)
}
