% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{check_upper_bound}
\alias{check_upper_bound}
\title{Check the upper value to be used meets criteria}
\usage{
check_upper_bound(upper)
}
\arguments{
\item{upper}{double The upper bound for the params}
}
\description{
Generates upper and lower bounds for all parameters
}
\examples{
\dontrun{
check_upper_bound(upper = 10)
}

}
