% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_objective}
\alias{fit_objective}
\title{Fit wagglefit model}
\usage{
fit_objective(x, mtype)
}
\arguments{
\item{x}{double foraging distances}

\item{mtype}{char defaults to "0" The model to use
0 = all, 1 = scout, 2 = recruit}
}
\description{
fits the wagglefit model using numerical optmization
}
