% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_bounds}
\alias{generate_bounds}
\title{Generates bounds for all parameters}
\usage{
generate_bounds()
}
\value{
bounds named list of upper and lower bounds
}
\description{
Generates upper and lower bounds for all parameters
}
\examples{
\dontrun{
generate_bounds()
}

}
