% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_bounds_all}
\alias{generate_bounds_all}
\title{Generates bounds for all parameters of the scout recruit superposition model}
\usage{
generate_bounds_all(upper)
}
\arguments{
\item{upper}{double The upper bound for the params}
}
\value{
bounds matrix of lower ([,1]) and upper ([,2]) bounds
}
\description{
Generates upper and lower bounds for all parameters
}
\examples{
\dontrun{
generate_bounds_all()
}

}
