% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_starting_estimates}
\alias{generate_starting_estimates}
\title{Generates starting estimates for all numerical optimisation of all parameters}
\usage{
generate_starting_estimates(bounds)
}
\arguments{
\item{bounds}{matrix of lower ([,1]) and upper ([,2]) bounds}
}
\value{
starting_estimates doubleArray of starting parameter estimates
}
\description{
Generates starting estimates for all parameters by taking a
random value between it's upper and lower bound.
}
\examples{
\dontrun{
generate_starting_estimates(generate_bounds())
}

}
