% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{generate_starting_ests_scout}
\alias{generate_starting_ests_scout}
\title{Generates starting estimates for all numerical optimisation of all parameters
of the scout model (can also be used for the recruit model)}
\usage{
generate_starting_ests_scout(distance, bounds, verbose = FALSE)
}
\arguments{
\item{distance}{doubleArray foraging distance to fit.}

\item{bounds}{matrix of lower ([,1]) and upper ([,2]) bounds}

\item{verbose}{Bool, to display function evaluations, defaults to FALSE}
}
\value{
starting_estimates doubleArray of starting parameter estimates
}
\description{
Generates starting estimates for all parameters by taking a
random value sampled from a normal distribution truncated at the upper and
lower bounds. Mean is given as the mid point between the upper and lower
bounds and the sd is set as 1.0
}
\examples{
\dontrun{
generate_starting_estimates_all(distance, generate_bounds())
}

}
