% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{get_bounds_collective}
\alias{get_bounds_collective}
\title{Generates bounds for the collective model}
\usage{
get_bounds_collective(upper)
}
\arguments{
\item{upper}{double The upper bound for the params}
}
\value{
bounds matrix of lower ([,1]) and upper ([,2]) bounds
}
\description{
Generates upper and lower bounds for all parameters
}
\examples{
\dontrun{
generate_bounds_all()
}

}
