% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{get_bounds_individual}
\alias{get_bounds_individual}
\title{Generates bounds for all parameters of the individual model}
\usage{
get_bounds_individual(upper)
}
\arguments{
\item{upper}{double The upper bound for the params}
}
\value{
bounds matrix of lower ([,1]) and upper ([,2]) bounds
}
\description{
Generates upper and lower bounds for all parameters
}
\examples{
\dontrun{
generate_bounds_all()
}

}
