% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_model_parameter_bounds}
\alias{get_model_parameter_bounds}
\title{Create model starting parameter bounds}
\usage{
get_model_parameter_bounds(mtype)
}
\arguments{
\item{mtype}{char defaults to "0" The model to use
0 = all, 1 = scout, 2 = recruit}
}
\description{
generates parameter bounds for numical optimisation
}
