% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_model_starting_estimates}
\alias{get_model_starting_estimates}
\title{Create model starting estimates}
\usage{
get_model_starting_estimates(mtype)
}
\arguments{
\item{mtype}{char defaults to "0" The model to use
0 = all, 1 = scout, 2 = recruit}
}
\description{
generates random starting values for numical optimisation
}
