% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{heaviside}
\alias{heaviside}
\title{Heaviside equation. Replace negatives with 0 and positives with 1.}
\usage{
heaviside(x)
}
\arguments{
\item{x}{double the parameter to convert with heaviside}
}
\description{
Heaviside equation. Replace negatives with 0 and positives with 1.
}
