% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{inverse_ccdf}
\alias{inverse_ccdf}
\title{Calculates the ccdf for given data}
\usage{
inverse_ccdf(x)
}
\arguments{
\item{x}{The data, e.g. foraging distance}
}
\value{
tibble of the sorted foraging distances and their cumulative
probability
}
\description{
Orders the data smallest to largest and calculates the
probability of sampling a value greater than or equal to a given value.
}
