% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglike_model_all}
\alias{loglike_model_all}
\title{Log-likelihood function for scout and recruit superposition}
\usage{
loglike_model_all(x, x_size, p, ls, ln, qn, a)
}
\arguments{
\item{x}{double* Pointer to array of foraging distances}

\item{x_size}{int Number of foraging distances}

\item{p}{double Proportion of scouts (0<=p<=1)}

\item{ls}{double Scout rate}

\item{ln}{double Recruit rate}

\item{qn}{double Quality}

\item{a}{double alpha value}
}
\description{
Log-likelihood function for scout and recruit superposition
}
