% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglike_model_all_new}
\alias{loglike_model_all_new}
\title{Log-likelihood function for scout and recruit superposition using
re-parameterised model}
\usage{
loglike_model_all_new(x, p, bs, br, as, ar)
}
\arguments{
\item{x}{double* Pointer to array of foraging distances}

\item{p}{double Proportion of scouts (0<=p<=1)}

\item{bs}{double scout rate}

\item{br}{double Recruit rate}

\item{as}{double scout alpha}

\item{ar}{double Recruit alpha}
}
\description{
Log-likelihood function for scout and recruit superposition using
re-parameterised model
}
