% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglike_model_individual}
\alias{loglike_model_individual}
\title{Log-likelihood function for individual foraging}
\usage{
loglike_model_individual(x, bs, as)
}
\arguments{
\item{x}{double* Pointer to array of foraging distances}

\item{bs}{double scout rate}

\item{as}{double scout alpha}
}
\description{
Log-likelihood function for individual foraging
}
