% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglike_model_recruit}
\alias{loglike_model_recruit}
\title{Log-likelihood function for recruits}
\usage{
loglike_model_recruit(x, x_size, ln, qn, a)
}
\arguments{
\item{x}{double* Pointer to array of foraging distances}

\item{x_size}{int Number of foraging distances}

\item{ln}{double Recruit rate}

\item{qn}{double Quality}

\item{a}{double alpha value}
}
\description{
Log-likelihood function for recruits
}
