% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglike_model_scout}
\alias{loglike_model_scout}
\title{Log-likelihood function for scouts}
\usage{
loglike_model_scout(x, x_size, ls, qn, a)
}
\arguments{
\item{x}{double* Pointer to array of foraging distances}

\item{x_size}{int Number of foraging distances}

\item{ls}{double Scout rate}

\item{qn}{double Quality}

\item{a}{double alpha value}
}
\description{
Log-likelihood function for scouts
}
