% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_ccdf_plot_data}
\alias{make_ccdf_plot_data}
\title{Creates data for plots}
\usage{
make_ccdf_plot_data(x, param_est, model)
}
\arguments{
\item{x}{doubleArray The foraging distance to plot}

\item{param_est}{doubleArray The paramater estimates of the optimisation}

\item{model}{characterArray The name of the model used}
}
\value{
tibble of model data and cumulative probability
}
\description{
Using the optimised paramater estimates, creates the data for
the ccdf plots
}
