% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_full_plot}
\alias{make_full_plot}
\title{Creates plots with model fits from optimised paramaters}
\usage{
make_full_plot(x, param_est)
}
\arguments{
\item{x}{The foraging distance to plot}

\item{param_est}{The paramater estimates of the optimisation}
}
\value{
ggplot plot of cumulative probability of foraging distances in Km
along with the model fits to this data.
}
\description{
Using the optimised paramater estimates, creates ccdf plots
}
