% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_results_tibble}
\alias{make_results_tibble}
\title{Converts results list to tibble}
\usage{
make_results_tibble(result)
}
\arguments{
\item{result}{namedList the results list}
}
\value{
tibble of model results
}
\description{
Converts the results list from running the models into a wide
format tibble for each model
}
