% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_likelihood_space}
\alias{map_likelihood_space}
\title{Calculate likelihoods for all parameters for a given dataset}
\usage{
map_likelihood_space(
  data,
  params,
  model_name = "collective",
  n = 1000,
  upper = 5,
  bounds = NULL
)
}
\arguments{
\item{params}{named list or tibble of optimised parameter values}

\item{model_name}{character array The name of the model to evaluate}
}
\value{
ggplotObj The likelihood plot for each parameter.
}
\description{
Calculate likelihoods for all parameters to show their
likelihood space for a given set of data and optimised model parameters. The
idea here is that once you have optimised your parameters you pass them in
here with the data and it shows the likelihood smace of each one wrt. the
others fixed at their optima.
}
\concept{utility}
