% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_verbose}
\alias{message_verbose}
\title{Wrapper for message}
\usage{
message_verbose(verbose = TRUE, ...)
}
\arguments{
\item{verbose}{Logical, defaults to `TRUE`. Should verbose processing
messages and warnings be returned.}

\item{...}{Additional arguments passed to `message`.}
}
\description{
A wrapper for `message` that only prints output when
`verbose = TRUE`.
}
\concept{utility}
