% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model}
\alias{model}
\title{Call desired model}
\usage{
model(x, mtype, ...)
}
\arguments{
\item{x}{double foraging distances}

\item{mtype}{char defaults to "0" The model to use
0 = all, 1 = scout, 2 = recruit}

\item{...}{parameters for the required model}
}
\value{
The requested distribution
}
\description{
Calls the requested distribution model (all, scout or recruit)
}
