% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_all}
\alias{model_all}
\title{Calculate distribution of recruit and scout dances combined}
\usage{
model_all(x, ...)
}
\arguments{
\item{x}{double foraging distances}

\item{...}{parameters for the required model}
}
\value{
The scout and recruit model distribution
}
\description{
Calculates superimposed distribution of recruit and scout
foraging distances
}
