% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_loglike}
\alias{model_loglike}
\title{Get log-likelihood of requested model}
\usage{
model_loglike(x, mtype = "0", ...)
}
\arguments{
\item{x}{double foraging distances}

\item{mtype}{char defaults to "0" The model to use
0 = all, 1 = scout, 2 = recruit}

\item{...}{parameters for the required model}
}
\value{
The requested distribution log-likelihood
}
\description{
Calculates log-likelihood of the distribution of the requested
models
}
