% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_number_from_model}
\alias{model_number_from_model}
\title{Get model number from model name}
\usage{
model_number_from_model(model)
}
\arguments{
\item{model}{characterArray The model to get the number for}
}
\description{
Converts the model name to its number for cpp. E.g. 'all' is 0
}
\concept{utility}
