% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_recruit}
\alias{model_recruit}
\title{Calculate distribution of recruit model dances}
\usage{
model_recruit(x, ...)
}
\arguments{
\item{x}{double foraging distances}

\item{...}{parameters for the required model}
}
\value{
The recruit model distribution
}
\description{
Calculates distribution of recruit foraging distances for
the model
}
