% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_scout}
\alias{model_scout}
\title{Calculate distribution of scout model dances}
\usage{
model_scout(x, ...)
}
\arguments{
\item{x}{double foraging distances}

\item{...}{parameters for the required model}
}
\value{
The scout model distribution
}
\description{
Calculates distribution of scout foraging distances for
the model
}
