% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{objective}
\alias{objective}
\title{Objective function for wagglefit model}
\usage{
objective(x, mtype, ...)
}
\arguments{
\item{x}{double foraging distances}

\item{mtype}{char defaults to "0" The model to use
0 = all, 1 = scout, 2 = recruit}

\item{...}{parameters for the required model}
}
\value{
negative log-likelihood for wagglefit model
}
\description{
wrapps `model_loglike` returning the negative log-likelihood
}
