% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimise_all}
\alias{optimise_all}
\title{Optimise function for scout and recruit superposition}
\usage{
optimise_all(x, params, lb, ub)
}
\arguments{
\item{x}{NumericVector Foraging distance}

\item{params}{NumericVector parameter estimates to run the model with}

\item{lb}{NumericVector, array of lower bounds for each paramater}

\item{ub}{NumericVector, array of upper bounds for each parameter}
}
\description{
Optimise function for scout and recruit superposition
}
