% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimize_all}
\alias{optimize_all}
\title{Optimize function for scout and recruit superposition}
\usage{
optimize_all(x, params, lb, ub)
}
\arguments{
\item{lb}{NumericVector, array of lower bounds for each paramater}

\item{ub}{NumericVector, array of upper bounds for each parameter}
}
\description{
Optimize function for scout and recruit superposition
}
