% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdf_model_all}
\alias{pdf_model_all}
\title{Model pdf function for scout and recruit superposition. Stores results in
given array (y)}
\usage{
pdf_model_all(x, y, p, bs, br, as, ar)
}
\arguments{
\item{x}{NumericVector foraging distances}

\item{y}{NumericVector storage array for the results}

\item{p}{double Proportion of scouts (0<=p<=1)}

\item{bs}{double scout rate}

\item{br}{double Recruit rate}

\item{as}{double scout alpha}

\item{ar}{double Recruit alpha}
}
\description{
Model pdf function for scout and recruit superposition. Stores results in
given array (y)
}
