% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdf_model_scout}
\alias{pdf_model_scout}
\title{Model pdf function for scout model. Stores results in given array (y)}
\usage{
pdf_model_scout(x, y, bs, as)
}
\arguments{
\item{x}{NumericVector foraging distances}

\item{y}{NumericVector storage array for the results}

\item{bs}{double scout rate}

\item{as}{double scout alpha}
}
\description{
Model pdf function for scout model. Stores results in given array (y)
}
