% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recruit_ccdf}
\alias{recruit_ccdf}
\title{Model CCDF function for recruits}
\usage{
recruit_ccdf(x, p, ln, qn, a, m)
}
\arguments{
\item{x}{double Foraging distance}

\item{p}{double Proportion of scouts (0<=p<=1)}

\item{ln}{double Recruit rate}

\item{qn}{double Quality}

\item{a}{double alpha value}

\item{m}{double Minimum foraging distance}
}
\value{
result double the ccdf for that foraging distance (x)
}
\description{
Model CCDF function for recruits
}
