% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{recruit_dist}
\alias{recruit_dist}
\title{Calculate distribution of recruit dances}
\usage{
recruit_dist(x, p, ln, qn, a)
}
\arguments{
\item{x}{double foraging distances}

\item{p}{double proportion of scouts}

\item{ln}{double the recruit rate}

\item{qn}{double quality}

\item{a}{double alpha value}
}
\value{
The scout distribution
}
\description{
Calculates distribution of recruit foraging distances
}
