% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recruit_dist_new}
\alias{recruit_dist_new}
\title{New model function for recruits}
\usage{
recruit_dist_new(x, m, br, ar)
}
\arguments{
\item{x}{double Foraging distance}

\item{m}{double Minimum foraging distance}

\item{br}{double Recruit rate}

\item{ar}{double Recruit alpha}
}
\description{
New model function for recruits
}
