% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run}
\alias{run}
\title{Run the full model on data}
\usage{
run(
  distance,
  model = "all",
  upper = 5,
  verbose_r = FALSE,
  verbose_cpp = FALSE,
  xtol = 0,
  iteration = c(1, 1)
)
}
\arguments{
\item{distance}{doubleArray The distance decoded from the waggle dance.}

\item{model}{characterArray The model to run (scout or all)}

\item{upper}{the upper limmit of the parameter bounds}

\item{verbose_r}{Bool, to display R function evaluations, defaults to FALSE}

\item{verbose_cpp}{Bool, to display C function evaluations, defaults to FALSE}

\item{xtol}{double, The absolute tolerance on function value. If 0 (default)
then default to nlopt default value.}

\item{iteration}{integerArray The number of times fit has been called out of
the number to be called. Defaults to c(1, 1)}
}
\description{
Run the full model on given data once.
}
