% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scout_dist}
\alias{scout_dist}
\title{Model function for scouts}
\usage{
scout_dist(x, m, p, ls, qn, a)
}
\arguments{
\item{x}{double Foraging distance}

\item{m}{double Minimum foraging distance}

\item{p}{double Proportion of scouts (0<=p<=1)}

\item{ls}{double Scout rate}

\item{qn}{double Quality}

\item{a}{double alpha value}
}
\description{
Model function for scouts
}
