% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scout_dist_new}
\alias{scout_dist_new}
\title{New model function for scouts}
\usage{
scout_dist_new(x, m, bs, as)
}
\arguments{
\item{x}{double Foraging distance}

\item{m}{double Minimum foraging distance}

\item{bs}{double scout rate}

\item{as}{double scout alpha}
}
\description{
New model function for scouts
}
