% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trunc_normal}
\alias{trunc_normal}
\title{Draws sample from a normal distribtion trucated between upr and lwr bounds.}
\usage{
trunc_normal(mean, sd, lwr, upr)
}
\arguments{
\item{mean}{double Mean for the normal distribution to sample}

\item{sd}{double Standard deviation for the normal distribution to sample}

\item{lwr}{double Lower bound samples must be >= to}

\item{upr}{double Upper bound samples must be <= to}
}
\value{
x double The sampled value
}
\description{
Samples from a normal distribution and then checks it falls
within given bounds.
}
\examples{
\dontrun{
trunc_normal(0, 0.5, -1, 1)
}

}
