% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_dist}
\alias{calc_dist}
\title{Caluclates distance from waggle dance}
\usage{
calc_dist(duration)
}
\arguments{
\item{duration}{double The duration of the waggle run in seconds.}
}
\value{
The distance indicated by the dance
}
\description{
Calculates distance in Km from dance duration, using a linear
relationship. Taken from Schürch et al (2019)
}
\details{
Calculate distance from waggle run duration
}
\examples{
\dontrun{
calc_dist(1.8)
}
}
\concept{utility}
