% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccdf_model_scout}
\alias{ccdf_model_scout}
\title{Model ccdf function for scout model. Stores results in given array (y)}
\usage{
ccdf_model_scout(x, y, ls, qn, a)
}
\arguments{
\item{x}{NumericVector foraging distances}

\item{y}{NumericVector storage array for the results}

\item{ls}{double Scout rate}

\item{qn}{double Quality}

\item{a}{double alpha value}
}
\description{
Model ccdf function for scout model. Stores results in given array (y)
}
