% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit}
\alias{fit}
\title{fit model to data}
\usage{
fit(distance, upper = 5)
}
\arguments{
\item{distance}{doubleArray The distance decoded from the waggle dance in
meters.}

\item{upper}{double The upper parameter bound. Defaults to 5.}
}
\value{
Not sure yet
}
\description{
Fits specified model to the data given using MLE.
}
\examples{
\dontrun{
fit(distance)
}

}
