% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit}
\alias{fit}
\title{fit either the scout and recruit superposition model or the scout model to
data}
\usage{
fit(
  distance,
  model = "all",
  upper = 5,
  iteration = c(1, 1),
  verbose_r = FALSE,
  verbose_cpp = FALSE,
  xtol = 0
)
}
\arguments{
\item{distance}{doubleArray The distance decoded from the waggle dance.}

\item{model}{characterArray The model to run (scout or all)}

\item{upper}{double The upper parameter bound. Defaults to 5.}

\item{iteration}{integerArray The number of times fit has been called out of
the number to be called. Defaults to c(1, 1)}

\item{verbose_r}{Bool, to display R function evaluations, defaults to FALSE}

\item{verbose_cpp}{Bool, to display C function evaluations, defaults to FALSE}

\item{xtol}{double, The absolute tolerance on function value. If 0 (default)
then default to nlopt default value.}
}
\value{
Not sure yet
}
\description{
Fits specified model to the data given using MLE.
}
\examples{
\dontrun{
fit_all(distance)
}

}
