% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_full_plot}
\alias{make_full_plot}
\title{Creates plots with model fits from optimised paramaters}
\usage{
make_full_plot(x, model_result_list)
}
\arguments{
\item{x}{The foraging distance to plot}

\item{model_result_list}{namedlist The return of `fit`, a list of parameter
estimates ($est) and the data name ($data_name).}
}
\value{
ggplot plot of cumulative probability of foraging distances in Km
along with the model fits to this data.
}
\description{
Using the optimised paramater estimates, creates ccdf plots
}
