% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_full_plot}
\alias{make_full_plot}
\title{Creates plots with model fits from optimised paramaters}
\usage{
make_full_plot(
  x,
  model_result_list,
  logit = TRUE,
  subplot_coords = c(0.5, 2.5, -7.8, -2.5)
)
}
\arguments{
\item{x}{The foraging distance to plot}

\item{model_result_list}{namedlist The return of `fit`, a list of parameter
estimates ($est) and the data name ($data_name).}

\item{logit}{Bool To log the data or not. Defaults to TRUE}

\item{subplot_coords}{doubleArray Coordinates for location of inner histogram}
}
\value{
ggplot plot of cumulative probability of foraging distances in Km
along with the model fits to this data.
}
\description{
Using the optimised paramater estimates, creates ccdf plots
}
