% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimise_all}
\alias{optimise_all}
\title{Optimise function for scout and recruit superposition}
\usage{
optimise_all(x, params, lb, ub, verbose = FALSE, xtol = 0)
}
\arguments{
\item{x}{NumericVector Foraging distance}

\item{params}{NumericVector parameter estimates to run the model with}

\item{lb}{NumericVector, array of lower bounds for each paramater}

\item{ub}{NumericVector, array of upper bounds for each parameter}

\item{verbose}{Bool, to display optimisation as it runs, defaults to FALSE}

\item{xtol}{double, The absolute tolerance on function value. If 0 (default)
then default to nlopt default value.}
}
\description{
Optimise function for scout and recruit superposition
}
