% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recruit_ccdf}
\alias{recruit_ccdf}
\title{Model CCDF for recruit distribution}
\usage{
recruit_ccdf(x, m, br, ar)
}
\arguments{
\item{x}{double Foraging distance}

\item{m}{double Minimum foraging distance}

\item{br}{double Recruit rate}

\item{ar}{double recruit alpha}
}
\value{
result double the ccdf for that foraging distance (x)
}
\description{
Model CCDF for recruit distribution
}
