% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recruit_dist}
\alias{recruit_dist}
\title{Model function for recruits}
\usage{
recruit_dist(x, m, p, ln, qn, a)
}
\arguments{
\item{x}{double Foraging distance}

\item{m}{double Minimum foraging distance}

\item{p}{double Proportion of scouts (0<=p<=1)}

\item{ln}{double Recruit rate}

\item{qn}{double Quality}

\item{a}{double alpha value}
}
\description{
Model function for recruits
}
