% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recruit_dist}
\alias{recruit_dist}
\title{Model function for recruits}
\usage{
recruit_dist(x, m, br, ar)
}
\arguments{
\item{x}{double Foraging distance}

\item{m}{double Minimum foraging distance}

\item{br}{double Recruit rate}

\item{ar}{double Recruit alpha}
}
\description{
Model function for recruits
}
