% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scout_ccdf}
\alias{scout_ccdf}
\title{Model CCDF for scout distribution}
\usage{
scout_ccdf(x, m, bs, as)
}
\arguments{
\item{x}{double Foraging distance}

\item{m}{double Minimum foraging distance}

\item{bs}{double Scout rate}

\item{as}{double scout alpha}
}
\value{
result double the ccdf for that foraging distance (x)
}
\description{
Model CCDF for scout distribution
}
