% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scout_dist}
\alias{scout_dist}
\title{Model function for scouts}
\usage{
scout_dist(x, m, bs, as)
}
\arguments{
\item{x}{double Foraging distance}

\item{m}{double Minimum foraging distance}

\item{bs}{double scout rate}

\item{as}{double scout alpha}
}
\description{
Model function for scouts
}
