% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{scout_dist}
\alias{scout_dist}
\title{Calculate distribution of scout dances}
\usage{
scout_dist(x, p, ls, qn, a)
}
\arguments{
\item{x}{double foraging distances}

\item{p}{double proportion of scouts}

\item{ls}{double scout rate}

\item{qn}{double quality}

\item{a}{double alpha value}
}
\value{
The scout distribution
}
\description{
Calculates distribution of scout foraging distances
}
